requirements.txt
pillow
requests
pysimplegui
